Add files via upload
Reverse Shell with a Trigger Mechanism

This project involves creating a reverse shell that grants access to a target computer upon triggering. The reverse shell is activated when the user clicks on an image, causing a pop-up to appear. Once triggered, the initiating computer gains remote access to the target system, enabling command execution and file access.

The project showcases skills in cybersecurity, ethical hacking, and Python programming. It demonstrates understanding of reverse shell techniques, network protocols, and exploit mechanisms, all built within a controlled and educational environment for learning purposes.

⚠️ Note: Reverse shell projects should only be used in ethical contexts, such as authorized penetration testing or cybersecurity education. Unauthorized use is illegal and unethical.